add powershell compatibility with powershell script with other mods
Create conditional logic that allows previous primary and secondary
screenshot software to include powershell.

Powershell script uses powershell.exe to call PrnSc button
and waits for image to load to clipboard after selection

(which is the reason why `start-process` was used instead of
    `call-process`)

Added functionality to create a subdirectory when saving
screenshots. Also allow for users to name the link saved
for readability

Signed-off-by: Andrew Berisha <andrew.berisha@sdggroup.com>